Added the resetState method. Added a ToDo List text file. Added controls
The resetState(int state) method allows you to reset a state. Currently,
its used if the player presses Backspace while in the PlayState - it
returns them to the menu, and resets the PlayState.

The todo list allows us to write our thoughts for what is left todo
programming wise in a file right within eclipse, rather than having to
open up google drive. Mostly for basic on the spot/simple stuff.

Added the options to change our "select" and "reset" controls.