Added Animations, Player, MapObject, Collisions (HUGE UPDATE - TONS OF BUGS)
This is a huge update.

I added in the MapObject class which all of our movment and collision
enabled players/enemies will extend.

I added the Player class which is obviously our player, and i got
collisions working (mostly).

Animations are also in. I created a basic sprite sheet with colored
rectangles to use. (Artists get in there and change those squares to
your real art!)

This was a huge update and its nowhere even close to being done or bug
free. The only reason i've committed this early is because at this point
the game is playable albeit extremely buggy. Now it's time for you guys
to give input on the game play mechanics, the character size compared to
the tiles, the art direction, etc so we can change values and mess
around with the aesthetics and game play now.

I would recommend that no one even try to touch/understand the
MapObject, Player, or even Tile and TileMap classes as i have a ton of
work to do commenting and cleaning/organizing them all to be at an
acceptable level as far as code readability goes. Everything is
definitely starting to get stringy and unorganized at this point, so ill
have to move methods/classes around until everything makes sense and is
where bit should be as far as functionality goes.

Main Bugs: Based on how i handled collisions and setting the player to
fall when no platform is underneath, you can see the player levitates
for a second until the platform is 1 pixel away, in which the character
then begins to fall. This is bad and bouncy and I'm not sure yet how
we'll fix it, however don't worry about this, there's so much more
cleanup i have to do before i even begin thinking about that.

Moving forward i would say: Artists start uploading Tile sheets and
player sprite sheets for animation. Level designers, feel free to start
building platforms in better locations.