Added pickups and abstracted fading
Abstracted the fading methods so that any GameState object can
implement the fading.

Added a pickups entities class. Spawns pickups randomly (currently set
to happen very often so that I could test it, but can be easily
changed). Detects the collision and disappears when character jumps on
pickup, but does not do anything yet.

Enjoy and let me know if i created any bugs in this commit; i had to do
some merging of files, so…..