Added camera scrolling left and right, Added more tiles left and right, Optimized loading of sprites
You can now travel left and right and the tile map will move while the
plaer stays centered.
I've simply duplicated the tile map we already have once to the right
and once to the left for testing.
With this many tiles things started lagging a lot so i optimized the way
in which each tile loaded their sprite image.