Fixed Tiles moving near the left edge of the screen. Re-added and fixed double jump. FPS works now
Tiles on the very left side of the screen would move on their own out of
place, so i fixed that.

Double jumps code wasnt working so i fixed that. In the end i think it
should be a pickup effect

The FPS should now display on all gamestates by pressing F1