Increased dimensions of game
NOTE: Now 800 x 800. All artwork has been updated to fit that size, so
no need to fix anything, but all new artwork should be made to fit this
size. (Also, note that we can revert this commit if we want to go back
to the original, but let us know early)

- Also updated the tileMap editor to support the 800 x 800 map (used it
to create level5, the test map currently in use)
- Updated all artwork
- Updated all text to be centered

FINAL NOTE: Everything should work, but if something broke with this
change, let me know asap so i can fix it.