Bug Fixes and Changes
- Abstracted some of the initializing done in Level1 state and Boss1
state into the play state (anything that all play states should be able
to do should now be in the play state)
- Issue #11 has been addressed
- Issue #10 has also been addressed (for the boss1 state. I think
there’s another reason for it appearing in level1 state). The player
can fall off the plane now.
- Issue #5 has also been addressed (assuming the freezing is caused by
window not being focused) If the game window loses focus, it will
automatically pause itself.
- Also added, but did not implement, a function that can allow us to
prevent player motion (this may be useful like at the end of level1
when the large stuka appears)