Added Bombs
Currently you just press B to spawn one in at a random x position.
There is a ConcurrentModificationException with the explosions for some
reason, I believe it is caused by multiple explosions happening at once,
so just dont spam and it should be ok.
This is just the first pass at bombs. I will improve them