Added a way to generate crash reports.
Upon crash (hopefully we don’t ever see this, but at least if the game
crashes, people will know that it did, and didn’t just freeze) a
message will appear telling the user of the crash and the location of
the auto generated file. A file with the stack trace is automatically
generated too.